Add Adobe Blank font repositories
We'll use Adobe Blank font and its build process as the basis of this
project. Adobe Blank 2 uses Format 13 cmap subtatble to make the font
very small in terms of download size, but it also lacks IE support
because of that. We'll use both version one and 2 to create 2 fonts, one
small one for modern browser and one "compat" font for legacy browsers.