Only load optimized font in Chromium & Ff
There have been several issues reported about browser compatibility with
IE, Edge (old non-chromium versions) and Safari.
See #10, #12, #13.

This seems to be because the native text renderers in Windows and some
versions of OSX/iOS don't seem to support cmap format 13, which would be
required for the optimized font to work.

Chrome and Firefox on the other hand seem to do their own text rendering
in a platform-independent way.

Switch to use the compatibility version of the font by default and
use some hacky CSS feature detection to only load the optimized font in
Chrome, Chromium-based Edge, and Firefox.